Fix user auth bugs for goal route
Fixed a bug in the model that caused goal text not to be created or
updated.

Fixed a bug in the controller that caused users to be able to edit
other users goals.

+++ backend/controllers/goalController.js
@@ -14,24 +14,29 @@ export const getGoals = asyncHandler(async (req, res) => {
 // @desc    Get goal
 // @route   GET /api/goals/:id
 // @access  Private
+// @TODO:   Delete this(?) I don't think this is useful to anyone
 export const getGoal = asyncHandler(async (req, res) => {
-  //TODO: Delete this(?)
   res.status(200).json({ message: `Get goal ${req.params.id}` })
 })
 
+////////////////////////////////////////////////////////////////////////////////
 // @desc    Set goal
 // @route   POST /api/goals
 // @access  Private
 export const setGoal = asyncHandler(async (req, res) => {
-  if (!req.body?.text) {
+  const text = req.body?.text
+  const user = req.user?.id
+  if (!text) {
     res.status(400)
     throw new Error("A 'text' field is required")
   }
 
-  const goal = await Goal.create({
-    text: req.body.text,
-  })
+  const goalObject = { text, user }
+  const goal = await Goal.create(goalObject)
 
+  console.log("------------------------------------------")
+  console.warn(goalObject)
+  console.log(goal)
   res.status(200).json({ goal })
 })
 
@@ -40,11 +45,19 @@ export const setGoal = asyncHandler(async (req, res) => {
 // @route   PUT /api/goals/:id
 // @access  Private
 export const updateGoal = asyncHandler(async (req, res) => {
-  // Get the goal from the db
+  const user = req.user?.id
+
+  // Get the goal from the db, throw an error if it doesn't exist
   const goal = await Goal.findById(req.params.id)
   if (!goal) {
     res.status(400)
-    throw new Eror("Goal not found")
+    throw new Error("Goal not found")
+  }
+
+  // Ensure the user matches
+  if (user !== goal.user.toString()) {
+    res.status(401)
+    throw new Error("User not authorized")
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -59,11 +72,21 @@ export const updateGoal = asyncHandler(async (req, res) => {
 // @route   DELETE /api/goals/:id
 // @access  Private
 export const deleteGoal = asyncHandler(async (req, res) => {
+  const user = req.user?.id
+
+  // Get the goal from the db, throw an error if it doesn't exist
   const goal = await Goal.findById(req.params.id)
   if (!goal) {
     res.status(400)
     throw new Error("Goal not found")
   }
+
+  // Ensure the user matches
+  if (user !== goal.user.toString()) {
+    res.status(401)
+    throw new Error("User not authorized")
+  }
+
   await goal.deleteOne()
   res.status(200).json({ id: req.params.id })
 })
